+++ __tests__/Bubble.test.js
@@ -3,11 +3,10 @@ import renderer from 'react-test-renderer';
 
 import { Bubble } from '../source/components';
 import { BubbleSkin } from '../source/skins/simple';
-import { CONTEXT } from './helpers/context';
 
 test('Bubble renders correctly', () => {
   const component = renderer.create(
-    <Bubble context={CONTEXT} skin={BubbleSkin} />
+    <Bubble skin={BubbleSkin} />
   );
 
   const tree = component.toJSON();
@@ -16,7 +15,7 @@ test('Bubble renders correctly', () => {
 
 test('Bubble renders isOpeningUpward={true}', () => {
   const component = renderer.create(
-    <Bubble isOpeningUpward context={CONTEXT} skin={BubbleSkin} />
+    <Bubble isOpeningUpward skin={BubbleSkin} />
   );
 
   const tree = component.toJSON();
@@ -25,7 +24,7 @@ test('Bubble renders isOpeningUpward={true}', () => {
 
 test('Bubble renders isTransparent={false}', () => {
   const component = renderer.create(
-    <Bubble isTransparent={false} context={CONTEXT} skin={BubbleSkin} />
+    <Bubble isTransparent={false} skin={BubbleSkin} />
   );
 
   const tree = component.toJSON();
@@ -34,7 +33,7 @@ test('Bubble renders isTransparent={false}', () => {
 
 test('Bubble renders isHidden={true}', () => {
   const component = renderer.create(
-    <Bubble isHidden context={CONTEXT} skin={BubbleSkin} />
+    <Bubble isHidden skin={BubbleSkin} />
   );
 
   const tree = component.toJSON();
@@ -43,7 +42,7 @@ test('Bubble renders isHidden={true}', () => {
 
 test('Bubble renders isFloating={true}', () => {
   const component = renderer.create(
-    <Bubble isFloating context={CONTEXT} skin={BubbleSkin} />
+    <Bubble isFloating skin={BubbleSkin} />
   );
 
   const tree = component.toJSON();