+++ __tests__/Button.test.js
@@ -3,11 +3,10 @@ import renderer from 'react-test-renderer';
 
 import { Button } from '../source/components';
 import { ButtonSkin } from '../source/skins/simple';
-import { CONTEXT } from './helpers/context';
 
 test('Button renders correctly', () => {
   const component = renderer.create(
-    <Button context={CONTEXT} skin={ButtonSkin} />
+    <Button skin={ButtonSkin} />
   );
 
   const tree = component.toJSON();
@@ -16,7 +15,7 @@ test('Button renders correctly', () => {
 
 test('Button renders with a label', () => {
   const component = renderer.create(
-    <Button label="send" context={CONTEXT} skin={ButtonSkin} />
+    <Button label="send" skin={ButtonSkin} />
   );
 
   const tree = component.toJSON();
@@ -25,7 +24,7 @@ test('Button renders with a label', () => {
 
 test('Button is disabled', () => {
   const component = renderer.create(
-    <Button disabled context={CONTEXT} skin={ButtonSkin} />
+    <Button disabled skin={ButtonSkin} />
   );
 
   const tree = component.toJSON();