+++ __tests__/Switch.test.js
@@ -3,12 +3,12 @@ import renderer from 'react-test-renderer';
 
 import { Checkbox } from '../source/components';
 import { SwitchSkin } from '../source/skins/simple';
-import { CONTEXT } from './helpers/context';
+import { IDENTIFIERS } from '../source/themes/API';
 
 test('Switch renders correctly', () => {
   const component = renderer.create(
     <Checkbox
-      context={CONTEXT}
+      themeId={IDENTIFIERS.SWITCH}
       skin={SwitchSkin}
     />
   );
@@ -20,8 +20,8 @@ test('Switch renders correctly', () => {
 test('Switch renders with label', () => {
   const component = renderer.create(
     <Checkbox
+      themeId={IDENTIFIERS.SWITCH}
       label="click here"
-      context={CONTEXT}
       skin={SwitchSkin}
     />
   );
@@ -33,8 +33,8 @@ test('Switch renders with label', () => {
 test('Switch is disabled', () => {
   const component = renderer.create(
     <Checkbox
+      themeId={IDENTIFIERS.SWITCH}
       disabled
-      context={CONTEXT}
       skin={SwitchSkin}
     />
   );
@@ -46,9 +46,9 @@ test('Switch is disabled', () => {
 test('Switch is disabled and renders a label', () => {
   const component = renderer.create(
     <Checkbox
+      themeId={IDENTIFIERS.SWITCH}
       disabled
       label="click here"
-      context={CONTEXT}
       skin={SwitchSkin}
     />
   );
@@ -60,8 +60,8 @@ test('Switch is disabled and renders a label', () => {
 test('Switch is checked', () => {
   const component = renderer.create(
     <Checkbox
+      themeId={IDENTIFIERS.SWITCH}
       checked
-      context={CONTEXT}
       skin={SwitchSkin}
     />
   );