+++ source/skins/simple/index.js
@@ -1,16 +1,16 @@
 // @flow
 // exports all skins as named exports
-export { default as AutocompleteSkin } from './AutocompleteSkin';
-export { default as BubbleSkin } from './BubbleSkin';
-export { default as ButtonSkin } from './ButtonSkin';
-export { default as CheckboxSkin } from './CheckboxSkin';
-export { default as FormFieldSkin } from './FormFieldSkin';
-export { default as InputSkin } from './InputSkin';
-export { default as ModalSkin } from './ModalSkin';
-export { default as OptionsSkin } from './OptionsSkin';
-export { default as RadioSkin } from './RadioSkin';
-export { default as SelectSkin } from './SelectSkin';
-export { default as SwitchSkin } from './SwitchSkin';
-export { default as TextAreaSkin } from './TextAreaSkin';
-export { default as TogglerSkin } from './TogglerSkin';
-export { default as TooltipSkin } from './TooltipSkin';
+export { AutocompleteSkin } from './AutocompleteSkin';
+export { BubbleSkin } from './BubbleSkin';
+export { ButtonSkin } from './ButtonSkin';
+export { CheckboxSkin } from './CheckboxSkin';
+export { FormFieldSkin } from './FormFieldSkin';
+export { InputSkin } from './InputSkin';
+export { ModalSkin } from './ModalSkin';
+export { OptionsSkin } from './OptionsSkin';
+export { RadioSkin } from './RadioSkin';
+export { SelectSkin } from './SelectSkin';
+export { SwitchSkin } from './SwitchSkin';
+export { TextAreaSkin } from './TextAreaSkin';
+export { TogglerSkin } from './TogglerSkin';
+export { TooltipSkin } from './TooltipSkin';